Widgets // Move desktop TimelineEntries to PZWidgetsKit.
- This allows them to be ViewModels.
