+++ WidgetExtension/Modules/TimelineProviders/Providers4Desktop/DualProfileWidgetProvider.swift
@@ -10,62 +10,6 @@ import PZInGameEventKit
 import PZWidgetsKit
 import WidgetKit
 
-// MARK: - DualProfileWidgetEntry
-
-@available(watchOS, unavailable)
-public struct DualProfileWidgetEntry: TimelineEntry, Sendable {
-    // MARK: Lifecycle
-
-    public init(
-        date: Date,
-        resultSlot1: Result<any DailyNoteProtocol, any Error>,
-        resultSlot2: Result<any DailyNoteProtocol, any Error>,
-        viewConfig: Config4DesktopProfileWidgets,
-        profileSlot1: PZProfileSendable?,
-        profileSlot2: PZProfileSendable?,
-        pilotAssetMap: [URL: SendableImagePtr]? = nil,
-        events: [OfficialFeed.FeedEvent]
-    ) {
-        self.date = date
-        self.resultSlot1 = resultSlot1
-        self.resultSlot2 = resultSlot2
-        self.viewConfig = viewConfig
-        self.profileSlot1 = profileSlot1
-        self.profileSlot2 = profileSlot2
-        self.events = events
-        self.pilotAssetMap = pilotAssetMap ?? [:]
-    }
-
-    // MARK: Public
-
-    public let date: Date
-    public let timestampOnCreation: Date = .now
-    public let resultSlot1: Result<any DailyNoteProtocol, any Error>
-    public let resultSlot2: Result<any DailyNoteProtocol, any Error>
-    public let viewConfig: Config4DesktopProfileWidgets
-    public let profileSlot1: PZProfileSendable?
-    public let profileSlot2: PZProfileSendable?
-    public let events: [OfficialFeed.FeedEvent]
-    public let pilotAssetMap: [URL: SendableImagePtr]
-
-    public var relevance: TimelineEntryRelevance? {
-        .init(score: Swift.max(countRelevance(resultSlot1), countRelevance(resultSlot2)))
-    }
-
-    public func countRelevance(_ result: Result<any DailyNoteProtocol, any Error>) -> Float {
-        switch result {
-        case let .success(data):
-            if data.staminaFullTimeOnFinish >= .now {
-                return 10
-            }
-            let stamina = data.staminaIntel
-            return 10 * Float(stamina.finished) / Float(stamina.all)
-        case .failure:
-            return 0
-        }
-    }
-}
-
 // MARK: - DualProfileWidgetProvider
 
 @available(watchOS, unavailable)
